SFDM-2514 Рарзработать метод GET, который возвращает всех пользователей с определенной ролью

See merge request products/eafdmmart/fdm-auth!23

+++ src/main/java/ru/beeline/fdmauth/repository/ProductRepository.java
@@ -13,7 +13,7 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
     @Query(value = "SELECT p.* FROM user_auth.product p " +
             "INNER JOIN user_auth.user_product upr ON p.id = upr.id_product " +
             "WHERE upr.id_profile = ?1", nativeQuery = true)
-    List<Product> getProductsByProfileId(Long idProfile);
+    List<Product> getProductsByProfileId(Integer idProfile);
 
     Product findAllByAlias(String alias);
 }
